Separated Display Api into its own header file
